guessing game - making casting

+++ 1_guessing_game/adivinhacao.c
@@ -46,7 +46,7 @@ int main() {
 
     tentativas++;
 
-    double pontosperdidos =   (chute - numerosecreto) / 2.0;
+    double pontosperdidos =   (chute - numerosecreto) / (double)2;
     pontos = pontos - pontosperdidos;
   }
